+++ src/registry/json_schema.rs
@@ -117,6 +117,7 @@ mod tests {
                 debug: 0,
                 quiet: false,
                 future: false,
+                allow_git_credentials: false,
                 command: Some(Commands::Registry(RegistryCommand {
                     command: RegistrySubCommand::JsonSchema(RegistryJsonSchemaArgs {
                         json_schema: json_schema_type.clone(),